% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering samples with controls}
\usage{
clustering(
  simMat_ = NULL,
  controls_ = NULL,
  thresh_ = NULL,
  smpl_graph = TRUE,
  disp_graph = TRUE
)
}
\arguments{
\item{simMat_}{Similarity matrix as a matrix object}

\item{controls_}{Indices of control samples in simMat_ as a string}

\item{thresh_}{Minimum for two samples being considered similar as a numeric}

\item{smpl_graph}{If sample graph must be output (default True)}

\item{disp_graph}{If dispersion graph must be output (default True)}
}
\value{
samples_table: samples with their assigned community (dense region)

cliques: cluster of samples

samples_graph: graph of samples colored by community

dispersion_graph: samples dispersed around contrl node
}
\description{
It needs some samples as controls and returns maximal cliques as clusters and highly interconnected regions as
communities.
}
\details{
TBD
}
\examples{
library(compaRe)

rm(list = ls())

# Step 1: Reading in similarity matrix

data(package = 'compaRe', list = c('simMat'))

# Step 2: Clustering

out_ = compaRe::clustering(simMat_ = as.matrix(simMat),
                           controls_ = "7,23,30,35,55,106,164,193,214,228,246,254,258,286,343,351,414,444,467,489,540",
                           thresh_ = NULL, smpl_graph = T, disp_graph = T)

# Step 3: Plotting dispersion graph

pdf(file = 'similarity_graph.pdf', width = 100, height = 100)
par(mar = c(0,0,0,0))
plot(out_$dispersion_graph, add = F, mark.groups = which(V(out_$dispersion_graph)$name \%in\% 'Control'), mark.col = 'mistyrose1', mark.expand = 2, mark.border = NA, directed = F)
graphics.off()

}
\seealso{
\code{\link{compare}} for measuring spatial similarity between two samples.
}
\author{
Morteza H. Chalabi, \url{mor.chalabi@gmail.com}
}
