% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\title{Compare data frames}
\usage{
library(compaRe)
compare_df(df1, df2, unique_id_df1, unique_id_df2)
}
\arguments{
\item{df1}{Old version of the dataset}

\item{df2}{Latest version of the dataset}

\item{unique_id_df1}{unique identifier in df1}

\item{unique_id_df2}{unique identifier in df2}
}
\description{
The \code{compare_df()} function compares two versions of the same data
}
\examples{
library(compaRe)

mtcars_v1 <- mtcars_v2 <- mtcars
mtcars_v1$id <- mtcars_v2$id <- rownames(mtcars)

mtcars_v2$am[mtcars_v2$am == 1] <- 2
mtcars_v2$wt[mtcars_v2$wt > 100] <- 105
mtcars_v2$cyl[mtcars_v2$cyl == 8] <- 7
mtcars_v2 <- mtcars_v2[1:30, ]

compare_df(df1 = mtcars1, df2 = mtcars2,
          unique_id_df1 = "id",
          unique_id_df2 = "id"
)
}
