% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_ce.R
\name{prob_ce}
\alias{prob_ce}
\title{Probability of composite binary endpoints}
\usage{
prob_ce(p_e1, p_e2, rho)
}
\arguments{
\item{p_e1}{numeric parameter, probability of the event E1}

\item{p_e2}{numeric parameter, probability of the event E2}

\item{rho}{numeric parameter, Pearson correlation between E1 and E2}
}
\value{
Returns the probability of the composite endpoint (E1 or E2).
}
\description{
This function calculates the probability of the composite of two events E1 and E2 (i.e., the union of the events E1 and E2).
This probability is calculated by means of the probabilities of each event and the correlation between them.
}
\details{
The input parameters represent the probability of the composite components and Pearson's correlation between the two components.
Note that Pearson's correlation takes values between two bounds that depend on the probabilities p0_e1 and p0_e2.
To calculate the correlation bounds you can use the R functions lower_corr and upper_corr, available in this package.
}
