% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_ce.R
\name{samplesize_ce}
\alias{samplesize_ce}
\title{Sample size for composite binary endpoints}
\usage{
samplesize_ce(
  p0_e1,
  p0_e2,
  type_e1,
  eff_e1,
  type_e2,
  eff_e2,
  effect_ce = "diff",
  rho,
  alpha = 0.05,
  beta = 0.2,
  unpooled = TRUE
)
}
\arguments{
\item{p0_e1}{numeric parameter, probability of occurrence E1 in the control group}

\item{p0_e2}{numeric parameter, probability of occurrence E2 in the control group}

\item{type_e1}{Effect measure used for the event E1  (type_e1 = "diff" for difference of proportions, type_e1 = "rr" for risk ratio, type_e1 = "or" for odds ratio)}

\item{eff_e1}{numeric parameter, anticipated effect for the composite component E1}

\item{type_e2}{Effect measure used for the event E2  (type_e2 = "diff" for difference of proportions, type_e2 = "rr" for risk ratio, type_e2 = "or" for odds ratio)}

\item{eff_e2}{numeric parameter, anticipated effect for the composite component E2}

\item{effect_ce}{Effect measure used for the composite endpoint (effect_ce = "diff" for difference of proportions, effect_ce = "rr" for risk ratio, effect_ce = "or" for odds ratio)}

\item{rho}{numeric parameter, Pearson's correlation between the two events E1 and E2}

\item{alpha}{Type I error}

\item{beta}{Type II error}

\item{unpooled}{Variance estimate used for the sample size calculation ("TRUE" for unpooled variance estimate, and "FALSE" for pooled variance estimate).}
}
\value{
Return the sample size for composite binary endpoints based on the anticipated values of the composite components
and the association between them in terms of Pearson's correlation.
}
\description{
This function calculates the required sample size for trials with a composite binary endpoint as primary endpoint.
The primary endpoint is assumed to be a composite binary endpoint formed by a combination of two events (E1 and E2).
The sample size is computed for evaluating differences between two groups  in terms of the risk difference, risk ratio or odds ratio.
The sample size is calculated on the basis of anticipated information of the composite components and the correlation between them.
}
\details{
The input parameters represent the probability of the composite components and Pearson's correlation between the two components.
Note that Pearson's correlation takes values between two bounds that depend on the probabilities p0_e1 and p0_e2.
To calculate the correlation bounds you can use the R functions lower_corr and upper_corr, available in this package.

 @references Bofill Roig, M., & Gómez Melis, G. (2019). A new approach for sizing trials with composite binary endpoints using anticipated marginal values and accounting for the correlation between components. Statistics in Medicine, 38(11), 1935–1956. https://doi.org/10.1002/sim.8092
}
