\name{sourceFile}
\alias{sourceFile}
\title{ Evaluate R Code from a File }
\description{
  Evaluate R code from a file within a local environment and return
  a list of objects of interest.
}
\usage{
sourceFile(filename, modelNames)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{ The name of a file containing R code. }
  \item{modelNames}{ The names of objects of interest that should be
     generated by the R code in the file and that we want returned. }
}
\value{
  A list containing one component for each name in \code{modelNames}.
}
\author{ Paul Murrell }
\seealso{
  \code{\link{compareFile}}
}
\examples{
modelNames <- c("id", "age", 
                "edu", "class", 
                "IndianMothers")
sourceFile(system.file("example", "student1.R", package="compare"),
           modelNames)
}
\keyword{utilities}
\keyword{logic}
