% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Comparing two samples by measuring their similarity}
\usage{
compare(smpl1 = NULL, smpl2 = NULL, merge_ = TRUE)
}
\arguments{
\item{smpl1}{First smaple as a matrix or data frame object}

\item{smpl2}{Second sample}

\item{merge_}{If two samples must be merged first (default True)}
}
\value{
Similarity value
}
\description{
It returns spatial similarity of two data sets with equal number of columns (variables) and any number
of rows (observations).
}
\details{
TBD
}
\examples{
require(compaRe)

rm(list = ls())

# Step 1: Reading in fcs files

data(package = 'compaRe', list = c('dataset1','dataset2'))

# Step 2: Transforming data

smpl1[ smpl1 < 0 ] = 0      # negative values in a fcs files are noise
smpl1 = log(smpl1+1)        # fcs data are logarithmically distributed
smpl2[ smpl2 < 0 ] = 0
smpl2 = log(smpl2+1)

# Step 3: Comparing (measuring similarity)

compaRe::compare(smpl1 = smpl1, smpl2 = smpl2, merge_ = TRUE)

}
\seealso{
\code{\link{clustering}} for clustering a similarity matrix generated by \code{compare}.
}
\author{
Morteza H. Chalabi, \url{mor.chalabi@gmail.com}
}
